moved NodeIterator to its own file
